Project Redirection - lots of breaking changes - read description (#11)
This PR introduced a lot of breaking changes but let me explain why to any future overseers.

The repository (before this change) was based around an article / blog system and was slowly transitioning to a user login system and slowly improving a poor backend.

This PR is the key turning point for the new idea that is a project showcase website for developers.

Therefore, this PR essentially moves the master branch to this 'new' area for refinement. It very much doesn't function as a production value and is considered still in development. However, its aim is to introduce fully tested classes and backend before working on the frontend.

Hope this is clear.